Menambahkan Authentication dan Authorization

+++ Controllers/EmployeeController.cs
@@ -21,6 +21,14 @@ namespace MVC75NET.Controllers
 
         public IActionResult Index()
         {
+            if (HttpContext.Session.GetString("email") == null)
+            {
+                return RedirectToAction("Unauthorized", "Error");
+            }
+            if (HttpContext.Session.GetString("role") != "Admin")
+            {
+                return RedirectToAction("Forbidden", "Error"); //RedirectToAction(Method,Controler)
+            }
             var tampil = empRepository.GetAll()
                 .Select(e => new EmployeeVM
                 {
@@ -38,6 +46,14 @@ namespace MVC75NET.Controllers
 
         public IActionResult Details(string id)
         {
+            if (HttpContext.Session.GetString("email") == null)
+            {
+                return RedirectToAction("Unauthorized", "Error");
+            }
+            if (HttpContext.Session.GetString("role") != "Admin")
+            {
+                return RedirectToAction("Forbidden", "Error"); //RedirectToAction(Method,Controler)
+            }
             var employees = empRepository.GetById(id);
             return View(new EmployeeVM
             {
@@ -53,11 +69,27 @@ namespace MVC75NET.Controllers
         }
         public IActionResult Create()
         {
+            if (HttpContext.Session.GetString("email") == null)
+            {
+                return RedirectToAction("Unauthorized", "Error");
+            }
+            if (HttpContext.Session.GetString("role") != "Admin")
+            {
+                return RedirectToAction("Forbidden", "Error"); //RedirectToAction(Method,Controler)
+            }
             return View();
         }
 
         public IActionResult Edit(string id)
         {
+            if (HttpContext.Session.GetString("email") == null)
+            {
+                return RedirectToAction("Unauthorized", "Error");
+            }
+            if (HttpContext.Session.GetString("role") != "Admin")
+            {
+                return RedirectToAction("Forbidden", "Error"); //RedirectToAction(Method,Controler)
+            }
             var employees = empRepository.GetById(id);
             return View(new EmployeeVM
             {
@@ -73,6 +105,14 @@ namespace MVC75NET.Controllers
         }
         public IActionResult Delete(string id)
         {
+            if (HttpContext.Session.GetString("email") == null)
+            {
+                return RedirectToAction("Unauthorized", "Error");
+            }
+            if (HttpContext.Session.GetString("role") != "Admin")
+            {
+                return RedirectToAction("Forbidden", "Error"); //RedirectToAction(Method,Controler)
+            }
             var employees = empRepository.GetById(id);
             return View(new EmployeeVM
             {
@@ -91,6 +131,14 @@ namespace MVC75NET.Controllers
         [ValidateAntiForgeryToken]
         public IActionResult Create(EmployeeVM employee)
         {
+            if (HttpContext.Session.GetString("email") == null)
+            {
+                return RedirectToAction("Unauthorized", "Error");
+            }
+            if (HttpContext.Session.GetString("role") != "Admin")
+            {
+                return RedirectToAction("Forbidden", "Error"); //RedirectToAction(Method,Controler)
+            }
             var result = empRepository.Insert(new Employee
             {
                 NIK = employee.NIK,
@@ -111,6 +159,14 @@ namespace MVC75NET.Controllers
         [ValidateAntiForgeryToken]
         public IActionResult Edit(EmployeeVM employee)
         {
+            if (HttpContext.Session.GetString("email") == null)
+            {
+                return RedirectToAction("Unauthorized", "Error");
+            }
+            if (HttpContext.Session.GetString("role") != "Admin")
+            {
+                return RedirectToAction("Forbidden", "Error"); //RedirectToAction(Method,Controler)
+            }
             var result = empRepository.Update(new Employee
             {
                 NIK = employee.NIK,
@@ -132,6 +188,14 @@ namespace MVC75NET.Controllers
         [ValidateAntiForgeryToken]
         public IActionResult Remove(string NIK)
         {
+            if (HttpContext.Session.GetString("email") == null)
+            {
+                return RedirectToAction("Unauthorized", "Error");
+            }
+            if (HttpContext.Session.GetString("role") != "Admin")
+            {
+                return RedirectToAction("Forbidden", "Error"); //RedirectToAction(Method,Controler)
+            }
             var result = empRepository.Delete(NIK);
             if (result > 0)
             {
